extract RobotContext references from swarm-scenario
