+++ src/swarm-scenario/Swarm/Game/Robot.hs
@@ -12,6 +12,12 @@
 -- A data type to represent robots.
 module Swarm.Game.Robot (
   -- * Robots data
+  RobotContextMember,
+  _robotID,
+  _robotLocation,
+  _robotContext,
+  _machine,
+  emptyActivityCount,
 
   -- * Robots
   RobotPhase (..),
@@ -24,12 +30,6 @@ module Swarm.Game.Robot (
   C.RobotUpdate (..),
 
   -- * Robot context
-  RobotContext,
-  defTypes,
-  defReqs,
-  defVals,
-  defStore,
-  emptyRobotContext,
   WalkabilityContext (..),
 
   -- ** Lenses
@@ -50,7 +50,6 @@ module Swarm.Game.Robot (
   inventoryHash,
   robotCapabilities,
   walkabilityContext,
-  robotContext,
   robotID,
   robotParentID,
   robotHeavy,
@@ -67,7 +66,6 @@ module Swarm.Game.Robot (
 
   -- ** Creation & instantiation
   mkRobot,
-  instantiateRobot,
 
   -- ** Query
   robotKnows,
@@ -82,11 +80,11 @@ module Swarm.Game.Robot (
 
 import Control.Applicative ((<|>))
 import Control.Lens hiding (Const, contains)
-import Data.Aeson qualified as Ae (FromJSON, Key, KeyValue, ToJSON (..), object, (.=))
+import Data.Aeson qualified as Ae (FromJSON, ToJSON (..))
 import Data.Hashable (hashWithSalt)
 import Data.Kind qualified
 import Data.Map (Map)
-import Data.Maybe (catMaybes, fromMaybe, isNothing)
+import Data.Maybe (isNothing)
 import Data.Sequence (Seq)
 import Data.Sequence qualified as Seq
 import Data.Set (Set)
@@ -94,17 +92,13 @@ import Data.Text (Text)
 import Data.Yaml (FromJSON (parseJSON), (.!=), (.:), (.:?))
 import GHC.Generics (Generic)
 import Linear
-import Servant.Docs (ToSample)
-import Servant.Docs qualified as SD
 import Swarm.Game.CESK qualified as C
 import Swarm.Game.Display (Display, curOrientation, defaultRobotDisplay, invisible)
 import Swarm.Game.Entity hiding (empty)
 import Swarm.Game.Location (Heading, Location, toDirection, toHeading)
-import Swarm.Game.Robot.Context
 import Swarm.Game.Universe
 import Swarm.Language.Capability (Capability)
 import Swarm.Language.Pipeline (ProcessedTerm)
-import Swarm.Language.Pipeline.QQ (tmQ)
 import Swarm.Language.Syntax (Const, Syntax)
 import Swarm.Language.Text.Markdown (Document)
 import Swarm.Language.Value as V
@@ -222,7 +216,6 @@ type instance RobotMachine 'ConcreteRobot = C.CESK
 
 type family RobotContextMember (phase :: RobotPhase) :: Data.Kind.Type
 type instance RobotContextMember 'TemplateRobot = ()
-type instance RobotContextMember 'ConcreteRobot = RobotContext
 
 -- | A value of type 'RobotR' is a record representing the state of a
 --   single robot.  The @f@ parameter is for tracking whether or not
@@ -256,7 +249,7 @@ deriving instance (Eq (RobotLocation phase), Eq (RobotID phase), Eq (RobotMachin
 -- See https://byorgey.wordpress.com/2021/09/17/automatically-updated-cached-views-with-lens/
 -- for the approach used here with lenses.
 
-makeLensesExcluding ['_robotCapabilities, '_equippedDevices, '_robotLog] ''RobotR
+makeLensesExcluding ['_robotCapabilities, '_equippedDevices, '_robotLog, '_robotContext] ''RobotR
 
 -- | A template robot, i.e. a template robot record without a unique ID number,
 --   and possibly without a location.
@@ -264,28 +257,6 @@ type TRobot = RobotR 'TemplateRobot
 
 -- | A concrete robot, with a unique ID number and a specific location.
 type Robot = RobotR 'ConcreteRobot
-
-instance ToSample Robot where
-  toSamples _ = SD.singleSample sampleBase
-   where
-    sampleBase :: Robot
-    sampleBase =
-      mkRobot
-        0
-        emptyRobotContext
-        Nothing
-        "base"
-        "The starting robot."
-        defaultCosmicLocation
-        zero
-        defaultRobotDisplay
-        (C.initMachine [tmQ| move |] mempty C.emptyStore)
-        []
-        []
-        False
-        False
-        mempty
-        0
 
 -- In theory we could make all these lenses over (RobotR phase), but
 -- that leads to lots of type ambiguity problems later.  In practice
@@ -364,29 +335,9 @@ robotOrientation = robotEntity . entityOrientation
 robotInventory :: Lens' Robot Inventory
 robotInventory = robotEntity . entityInventory
 
--- | The robot's context.
-robotContext :: Lens' Robot RobotContext
-
 -- | The (unique) ID number of the robot.  This is only a Getter since
 --   the robot ID is immutable.
 robotID :: Getter Robot RID
-
--- | Instantiate a robot template to make it into a concrete robot, by
---    providing a robot ID. Concrete robots also require a location;
---    if the robot template didn't have a location already, just set
---    the location to (0,0) by default.  If you want a different location,
---    set it via 'trobotLocation' before calling 'instantiateRobot'.
---
--- If a machine is not supplied (i.e. 'Nothing'), will fallback to any
--- program specified in the template robot.
-instantiateRobot :: Maybe C.CESK -> RID -> TRobot -> Robot
-instantiateRobot maybeMachine i r =
-  r
-    { _robotID = i
-    , _robotLocation = fromMaybe defaultCosmicLocation $ _robotLocation r
-    , _machine = fromMaybe (mkMachine $ _machine r) maybeMachine
-    , _robotContext = emptyRobotContext
-    }
 
 -- | The ID number of the robot's parent, that is, the robot that
 --   built (or most recently reprogrammed) this robot, if there is
@@ -483,10 +434,6 @@ walkabilityContext :: Getter Robot WalkabilityContext
 walkabilityContext = to $
   \x -> WalkabilityContext (_robotCapabilities x) (_unwalkableEntities x)
 
-mkMachine :: Maybe ProcessedTerm -> C.CESK
-mkMachine Nothing = C.Out VUnit C.emptyStore []
-mkMachine (Just pt) = C.initMachine pt mempty C.emptyStore
-
 -- | A general function for creating robots.
 mkRobot ::
   -- | ID number of the robot.
@@ -576,41 +523,6 @@ instance FromJSONE EntityMap TRobot where
       <*> liftE (v .:? "heavy" .!= False)
       <*> liftE (v .:? "unwalkable" ..!= mempty)
       <*> pure 0
-
-(.=?) :: (Ae.KeyValue a, Ae.ToJSON v, Eq v) => Ae.Key -> v -> v -> Maybe a
-(.=?) n v defaultVal = if defaultVal /= v then Just $ n Ae..= v else Nothing
-
-(.==) :: (Ae.KeyValue a, Ae.ToJSON v) => Ae.Key -> v -> Maybe a
-(.==) n v = Just $ n Ae..= v
-
-instance Ae.ToJSON Robot where
-  toJSON r =
-    Ae.object $
-      catMaybes
-        [ "id" .== (r ^. robotID)
-        , "name" .== (r ^. robotEntity . entityDisplay)
-        , "description" .=? (r ^. robotEntity . entityDescription) $ mempty
-        , "loc" .== (r ^. robotLocation)
-        , "dir" .=? (r ^. robotEntity . entityOrientation) $ zero
-        , "display" .=? (r ^. robotDisplay) $ (defaultRobotDisplay & invisible .~ sys)
-        , "program" .== (r ^. machine)
-        , "devices" .=? (map (^. _2 . entityName) . elems $ r ^. equippedDevices) $ []
-        , "inventory" .=? (map (_2 %~ view entityName) . elems $ r ^. robotInventory) $ []
-        , "system" .=? sys $ False
-        , "heavy" .=? (r ^. robotHeavy) $ False
-        , "log" .=? (r ^. robotLog) $ mempty
-        , -- debug
-          "capabilities" .=? (r ^. robotCapabilities) $ mempty
-        , "logUpdated" .=? (r ^. robotLogUpdated) $ False
-        , "context" .=? (r ^. robotContext) $ emptyRobotContext
-        , "parent" .=? (r ^. robotParentID) $ Nothing
-        , "createdAt" .=? (r ^. robotCreatedAt) $ 0
-        , "selfDestruct" .=? (r ^. selfDestruct) $ False
-        , "activity" .=? (r ^. activityCounts) $ emptyActivityCount
-        , "runningAtomic" .=? (r ^. runningAtomic) $ False
-        ]
-   where
-    sys = r ^. systemRobot
 
 -- | Is the robot actively in the middle of a computation?
 isActive :: Robot -> Bool
